Elaborate on the constraints example
The constraints example was a bit hard to understand without explanation on how the functions work. Therefore I added some explanation. Moreover, I believe the conjunction ought to be an OrConjunction, as it seems to be impossible to satisfy several equality constraints on the the same parameter. I tried it and the sampled configurations include values for coef0 and gamma only with the OrConjuction.