Check new neighbors deterministically (#195)
* Check new neighbors deterministically

One-exchange neighborhood iterator rigorously checks some of the
generated configurations. The check is performed with the probability 5
%, as decided by `np.random`.

Make the check deterministic (using an explicitly seeded
`np.random.RandomState`) instead of relying on the default numpy random
state, which need not be seeded.

* Make a test consistent with the current behavior

of `get_one_exchange_neighborhood`.

Update the hard-coded expected values.