change_hp_value: Do not disable a `||`-conditioned HP too eagerly
When searching the hierarchy of hyperparameters (HPs),
every time a HP is activated or deactivated,
its children need to be re-considered for activation or deactivation.
If there is no cycle in the condition graph,
this process will terminate and converge.

Before this commit, each HP was only considered at most once.
This suffices if there is no `OrConjunction` condition:
as soon as one of the parents of a HP is inactive,
the HP is necessarily inactive.