fix: Memory leak with UniformIntegerHyperparameter
When querying a large range for a UniformIntegerHyperparameter with a
small std.deviation and log scale, this could cause an infinite loop as
the reachable neighbors would be quickly exhausted, yet rejection
sampling will continue sampling until some arbitrary termination
criterion. Why this was causing a memory leak, I'm not entirely sure.

The solution now is that is we have seen a sampled value before, we
simply take the one "next to it".